+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/PostedPictureFront.js":"1","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/Homepage.js":"2","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/PostedPictureBack.js":"3","/home/afafibrahimi/Desktop/Portafolio/ig/src/reportWebVitals.js":"4","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/Navbar-items.js":"5","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/SelectPictureFront.js":"6","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/SelectPictureBack.js":"7","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/FilterPictureBack.js":"8","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/FilterPictureFront.js":"9","/home/afafibrahimi/Desktop/Portafolio/ig/src/index.js":"10","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/Footer.js":"11","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/NavBar.js":"12","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/FlipCard.js":"13"},{"size":887,"mtime":1608027323055,"results":"14","hashOfConfig":"15"},{"size":985,"mtime":1607695950627,"results":"16","hashOfConfig":"15"},{"size":1061,"mtime":1607705353998,"results":"17","hashOfConfig":"15"},{"size":362,"mtime":1606749743851,"results":"18","hashOfConfig":"15"},{"size":228,"mtime":1607508732456,"results":"19","hashOfConfig":"15"},{"size":1443,"mtime":1607705517354,"results":"20","hashOfConfig":"15"},{"size":1653,"mtime":1607705508158,"results":"21","hashOfConfig":"15"},{"size":785,"mtime":1608027323047,"results":"22","hashOfConfig":"15"},{"size":771,"mtime":1607705355746,"results":"23","hashOfConfig":"15"},{"size":500,"mtime":1606751706312,"results":"24","hashOfConfig":"15"},{"size":1098,"mtime":1607720475931,"results":"25","hashOfConfig":"15"},{"size":653,"mtime":1607529887138,"results":"26","hashOfConfig":"15"},{"size":522,"mtime":1608027323079,"results":"27","hashOfConfig":"15"},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1e5ippq",{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"39"},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"39"},{"filePath":"50","messages":"51","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"52","messages":"53","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"54","messages":"55","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/PostedPictureFront.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/Homepage.js",[],["56","57"],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/PostedPictureBack.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/reportWebVitals.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/Navbar-items.js",[],["58","59"],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/SelectPictureFront.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/SelectPictureBack.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/FilterPictureBack.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/CardLayouts/FilterPictureFront.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/index.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/Footer.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/NavBar.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/FlipCard.js",[],{"ruleId":"60","replacedBy":"61"},{"ruleId":"62","replacedBy":"63"},{"ruleId":"60","replacedBy":"64"},{"ruleId":"62","replacedBy":"65"},"no-native-reassign",["66"],"no-negated-in-lhs",["67"],["66"],["67"],"no-global-assign","no-unsafe-negation"]
+[{"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/Navbar/NavBar.js":"1","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/Footer/Footer.js":"2","/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/FlipCard/FlipCard.js":"3"},{"size":653,"mtime":1614609104115,"results":"4","hashOfConfig":"5"},{"size":1098,"mtime":1614609139624,"results":"6","hashOfConfig":"5"},{"size":522,"mtime":1614609184801,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1e5ippq",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/Navbar/NavBar.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/Footer/Footer.js",[],"/home/afafibrahimi/Desktop/Portafolio/ig/src/Components/FlipCard/FlipCard.js",[]]
