+++ src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
-import Footer from "./Components/Footer";
+import Footer from "./Components/Footer/Footer";
 import HomePage from "./Components/Homepage";
-import NavBar from "./Components/NavBar";
+import NavBar from "./Components/Navbar/NavBar";
 
 function App() {
   return (
